src/drawhello/requirements.txt
azure-storage-blob
azure-storage-queue
python-dotenv
streamlit
streamlit_drawable_canvas
requests